+++ Dumped/Assembly-CSharp/Gluon/TimeAttackRanking.cs
@@ -7,7 +7,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
 
-// Image 55: Assembly-CSharp.dll - Assembly: Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null
+// Image 58: Assembly-CSharp.dll - Assembly: Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null
 
 namespace Gluon
 {
@@ -28,7 +28,7 @@ namespace Gluon
 			public int eventId;
 			public ulong viewerId;
 			public int questId;
-			public int clearTime;
+			public float clearTime;
 			public int startTime;
 			public int endTime;
 			public string partyHash;
